Converts destination duration to float

+++ Characters/NPCs/Destination.cs
@@ -3,7 +3,7 @@ namespace Lerocia.Characters.NPCs {
 
   public class Destination {
     public Vector3 Position;
-    public int Duration;
+    public float Duration;
 
     public Destination(Vector3 position, int duration) {
       Position = position;
